requirements.txt
streamlit==1.37.*
pandas==2.2.*
plotly==5.23.*
openpyxl==3.1.*
reportlab==4.2.*        # for PDF export
python-docx==1.1.*      # for DOCX export
pydantic==2.8.*         # session schema
docxtpl==0.17.*         # DOCX templating
docx2pdf        # DOCX to PDF conversion
